RP-809: restore instantiation of script object for each pipeline run
This implementation is a bit unusual as it takes a script object and uses it as a prototype to create new script objects. Eventually, I'd  prefer to restore  the implementation that created  the Pipeline with the script classes and configuration, rather than script objects.  However, this will take more intensive refactoring and testing.